backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
aiobotocore
pytest
pytest-asyncio
greenlet
requests
slowapi
